Move example code to a new module
It was moved the ble example code to a new module named
"custom_ble_manager". The respective source .c file and header .h were
created.

The functionality is the same as the "ble_app_template" from the nordic
example, it is just a reorganization.

The nrf52840 code once uploaded to the development kit can be tested
with the application nrfConnect installed on an Android phone.
